refactor: extract settings window to src/ui/settings/
- Create ui/settings/window.rs with ~350 lines of settings UI code
- Add OnSettingsClose callback type for dependency decoupling
- Move ensure_hotkey_menu, open_settings_window, close_settings_window
- main.rs reduced from 1485 to 1116 lines (-369 lines)
