+++ src/ui/settings/mod.rs
@@ -1,0 +1,7 @@
+//! Settings window module.
+//!
+//! This module contains the settings window and its controls.
+
+pub mod window;
+
+pub use window::{close_settings_window, ensure_hotkey_menu, open_settings_window, OnSettingsClose};
